move stuff around and add comments
